Bugfix!  Wrong loop index variable.
I was testing clang-tidy and found this.  The old code wouldn't have been likely to be run or cause too much trouble -- it's in a fallback allocator's error path for specific values of n -- but a double-free (attempt to free released memory) could occur as out[i] could be freed multiple times.  The real index here is n, not i.